Rename app module to sample #22 tests removed sample package renamed Dangerfile path fixed
