+++ Dangerfile
@@ -18,7 +18,7 @@ end
 
 # Notify of outdated dependencies
 dependencyUpdatesHeader = "The following dependencies have later milestone versions:"
-dependencyReportsFile = "app/build/dependencyUpdates/report.txt"
+dependencyReportsFile = "sample/build/dependencyUpdates/report.txt"
 
 # Due to the formatting of this output file, we first check if there are any dependencies
 # by looking for a `->` arrow, then we check for the relevant headers. We do this to handle a case
