+++ sample/src/main/res/values/themes.xml
@@ -1,7 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-
-    <style name="Theme.PermissionsExample" parent="android:Theme.Material.Light.NoActionBar">
-        <item name="android:statusBarColor">@color/purple_700</item>
-    </style>
+    <style name="Theme.PermissionsExample" parent="android:Theme.Material.Light.NoActionBar" />
 </resources>
